+++ firestore-genai-chatbot/CHANGELOG.md
@@ -1,6 +1,37 @@
+## Version 0.0.10
+
+- Add docs on regional support for Gemini APIs
+
+## Version 0.0.9
+
+- fix typo in documentation
+
+## Version 0.0.8
+
+- Update documentation
+
+## Version 0.0.7
+
+- Update documentation
+- Update extensions display name
+
+## Version 0.0.6
+
+- Make model a string param, to allow for future changes to model names.
+
+- Add maxOutputTokens parameter
+
+- Update documentation
+
+- Add safety threshold params
+
+## Version 0.0.5
+
+- Add Vertex AI provider
+
 ## Version 0.0.4
 
-- Allow Application Default Credentials Authorization
+- Fix context parameter
 
 ## Version 0.0.3